[#83] SideBar 컴포넌트 구조 수정 (#93)
* 🚚Rename: SideBar, SideBarContent 컴포넌트 파일 위치 변경

* ♻️Refactor: 파일 경로가 변경됨에 따라 import문 수정

* ♻️Refactor: SideBarContent 컴포넌트 내용 구조 분해
